+++ R/SDistribution_MultivariateNormal.R
@@ -1,4 +1,4 @@
-  #' @include SetInterval_SpecialSet.R ParameterSet.R
+
 #-------------------------------------------------------------
 #  Distribution Documentation
 #-------------------------------------------------------------
@@ -90,14 +90,15 @@ MultivariateNormal$set("public","setParameterValue",function(..., lst = NULL, er
   if(is.null(lst))
     lst <- list(...)
   if(!is.null(lst$cov)){
-    if(any(dim(lst$cov) != c(self$getParameterValue("K"), self$getParameterValue("K"))))
-      lst$cov <- matrix(lst$cov, nrow = self$getParameterValue("K"), ncol = self$getParameterValue("K"))
+    if(any(dim(lst$cov) != c(self$getParameterValue("K"), self$getParameterValue("K"))) |
+       length(lst$cov) != self$getParameterValue("K")^2)
+      lst$cov <- suppressWarnings(matrix(lst$cov, nrow = self$getParameterValue("K"), ncol = self$getParameterValue("K")))
     lst$cov <- as.numeric(lst$cov)
   }
   if(!is.null(lst$prec)){
     if(any(dim(lst$prec) != c(self$getParameterValue("K"), self$getParameterValue("K"))) |
-       length(dim) != self$getParameterValue("K")^2)
-      lst$prec <- matrix(lst$prec, nrow = self$getParameterValue("K"), ncol = self$getParameterValue("K"))
+       length(lst$prec) != self$getParameterValue("K")^2)
+      lst$prec <- suppressWarnings(matrix(lst$prec, nrow = self$getParameterValue("K"), ncol = self$getParameterValue("K")))
     lst$prec <- as.numeric(lst$prec)
   }
   if(!is.null(lst$mean)){
@@ -173,8 +174,8 @@ MultivariateNormal$set("public","initialize",function(mean = rep(0,2), cov = c(1
   }
 
   super$initialize(decorators = decorators, pdf = pdf, rand = rand,
-                   support = Reals$new(dim = length(mean)),
-                   symmetric = FALSE,type = Reals$new(dim = "K"),
+                   support = setpower(Reals$new(), length(mean)),
+                   symmetric = FALSE,type = setpower(Reals$new(), length(mean)),
                    valueSupport = "continuous",
                    variateForm = "multivariate")
   invisible(self)