+++ R/SDistribution_Weibull.R
@@ -1,4 +1,4 @@
-#' @include SetInterval_SpecialSet.R ParameterSet.R
+
 #-------------------------------------------------------------
 # Weibull Distribution Documentation
 #-------------------------------------------------------------