+++ tests/testthat/test-DistMultivariateNormal.R
@@ -1,6 +1,6 @@
 library(testthat)
 
-context("Multivariat Normal distribution")
+context("Multivariate Normal distribution")
 
 test_that("constructor",{
   expect_equal(MultivariateNormal$new()$strprint(), "MultiNorm(mean = c(0, 0), cov = c(1, 0, 0, 1))")
@@ -8,7 +8,7 @@ test_that("constructor",{
   expect_silent(MultivariateNormal$new(mean = c(0,0,0), prec = c(1,0,0,0,1,0,0,0,1)))
   expect_message(MultivariateNormal$new(mean = c(0,0,0), prec = c(1,0,0,0,1,0,0,0,1), verbose = T))
   expect_error(MultivariateNormal$new(mean = 5),"Length of mean")
-  expect_warning(MultivariateNormal$new(mean = c(2,3), cov = matrix(c(1,3,4))))
+  expect_silent(MultivariateNormal$new(mean = c(2,3), cov = matrix(c(1,3,4))))
 })
 
 test_that("represent",{
@@ -25,8 +25,8 @@ test_that("parameters", {
   expect_equal(mvn$getParameterValue("prec"), matrix(c(1,0,0,0,1,0,0,0,1),nrow=3))
   expect_equal(mvn$setParameterValue(lst = list(prec = c(1,0,0,0,1,0,0,0,1)))$getParameterValue("prec"),
                matrix(c(1,0,0,0,1,0,0,0,1),nrow = 3))
-  expect_warning(expect_equal(mvn$setParameterValue(lst = list(prec = c(1,0,0,0,1,0,0,0,1,1)))$getParameterValue("prec"),
-               matrix(c(1,0,0,0,1,0,0,0,1),nrow = 3)))
+  expect_equal(mvn$setParameterValue(lst = list(prec = c(1,0,0,0,1,0,0,0,1,1)))$getParameterValue("prec"),
+               matrix(c(1,0,0,0,1,0,0,0,1),nrow = 3))
   expect_equal(mvn$setParameterValue(lst = list(mean = c(1)))$getParameterValue("mean"),
                c(1,1,1))
   expect_equal(mvn$setParameterValue(lst = list(mean = c(1,2,3,4)))$getParameterValue("mean"),
@@ -37,10 +37,10 @@ test_that("properties & traits",{
   expect_equal(mvn$valueSupport(), "continuous")
   expect_equal(mvn$variateForm(), "multivariate")
   expect_equal(mvn$symmetry(), "asymmetric")
-  expect_equal(mvn$inf(), -Inf)
-  expect_equal(mvn$sup(), Inf)
-  expect_equal(mvn$dmin(), -Inf)
-  expect_equal(mvn$dmax(), Inf)
+  expect_equal(mvn$inf(), Tuple$new(-Inf, -Inf, -Inf))
+  expect_equal(mvn$sup(), Tuple$new(Inf, Inf, Inf))
+  expect_equal(mvn$dmin(), Tuple$new(-Inf, -Inf, -Inf))
+  expect_equal(mvn$dmax(), Tuple$new(Inf, Inf, Inf))
   expect_equal(mvn$kurtosisType(), NA)
   expect_equal(mvn$skewnessType(), NA)
 })
@@ -71,4 +71,3 @@ test_that("statistics",{
   expect_equal(signif(mvn$pdf(1:2,2:3,3:4),3), c(2.366e-07, 7.835e-06))
   expect_equal(dim(mvn$rand(10)),c(10,3))
 })
-