+++ NEWS.md
@@ -6,6 +6,8 @@
 - Added checks to assert all packages are installed when required by distributions
 - Deprecated `package` field in distributions, now called `packages`
 - `packages` lists all packages required to be installed for a distribution, if `NULL` only distr6 required
+- Abstracted SetInterval and related classes to the `set6` package
+- Deprecated `listSpecialSets`, use `set6::listSpecialSets` instead
 
 # distr6 1.3.3